pyproject.toml
[tool.poetry]
name = "Copilot"
version = "1.1.0"
description = ""
authors = ["Michael S."]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "<3.14,>=3.11"
openai = "^1.1.1"
aiohttp = "^3.8.4"
aiohttp-rpc = "^1.3.3"
setuptools = "^80.9.0"
telethon = "^1.36.0"
tiktoken = "^0.10.0"
duckduckgo-search = "^8.0.0"
beautifulsoup4 = "^4.3.2"
newspaper3k = "^0.2.8"
numpy = "^2.3.2"
pydantic = "^2.8.2"
aiohttp-cors = "^0.8.1"
lxml = {extras = ["html-clean"], version = "^5.2.2"}
html2text = "^2025.4.15"
pandas = "^2.2.3"
openpyxl = "^3.1.5"
python-magic = "^0.4.27"
jinja2 = "^3.1.5"
pyinstaller = "^6.12.0"
aiofiles = "^24.1.0"
docling = "^2.36.1"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
ruff = "^0.9.10"

[tool.poetry.group.webview]
optional = true

[tool.poetry.group.webview.dependencies]
pywebview = {extras = ["gtk"], version = "^5.4"}
pygobject = "^3.50.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
